+++ docs/conf.py
@@ -18,11 +18,11 @@
 # -- Project information -----------------------------------------------------
 
 project = 'twill'
-copyright = '2022, C. Titus Brown, Ben R. Taylor et al'
-author = 'C. Titus Brown, Ben R. Taylor et al'
+copyright = '2022, C. Titus Brown, Ben R. Taylor, Christoph Zwerschke et al'
+author = 'C. Titus Brown, Ben R. Taylor, Christoph Zwerschke et al'
 
 # The full version, including alpha/beta/rc tags
-version = release = '3.0.3'
+version = release = '3.1.0'
 
 
 # -- General configuration ---------------------------------------------------